+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 <!-- markdownlint-disable MD041 -->
 
+## 0.2.0
+
+- Variables have been renamed to be consistent across all methods
+- **BREAKING:** getVideos() and getLiveVideos()
+  - renamed `lang` to `languages`
+  - renamed `sort` to `videoSort`
+  - renamed `status` to `videoStatus`
+  - renamed `type` to `videoType`
+- **BREAKING:** getVideos() renamed `topicId` to `topic`
+- **BREAKING:** getChannels()
+  - renamed `sort` to `channelSort`
+  - renamed `lang` to `languages`
+- **BREAKING:** getChannelVideos(), getVTuberClips(), and getVideosRelatedToChannel() renamed `lang` to `languages`
+- **BREAKING:** searchVideos() and searchComments() renamed `vch` to `videoChannels`
+
 ## 0.1.2+1
 
 - Initial release